requirements.txt
Flask
Flask_sqlalchemy
Flask_cors
python-dotenv
requests
gunicorn
psycopg2-binary
Flask-Migrate
Werkzeug